Code restructuring
This commit has two changes: (1) It moves ResourceTracker into a new
optimizer subdirectory. (2) It separates EngineFim (non-templated)
from EngineFimBase (templated) to be consistent with the way Fim and
FimBase work. This also reduces the number of imports needed in many
files.
