+++ packages/fim-browser/src/engine/BrowserEngineFim.ts
@@ -8,10 +8,10 @@ import { CoreBrowserCanvasWebGL } from '../core/CoreBrowserCanvasWebGL';
 import { CoreBrowserOffscreenCanvas2D } from '../core/CoreBrowserOffscreenCanvas2D';
 import { CoreBrowserOffscreenCanvasWebGL } from '../core/CoreBrowserOffscreenCanvasWebGL';
 import { FimDimensions, FimImageOptions } from '@leosingleton/fim';
-import { CoreCanvas2D, CoreCanvasOptions, CoreCanvasWebGL, EngineFim, EngineShader } from '@leosingleton/fim/internals';
+import { CoreCanvas2D, CoreCanvasOptions, CoreCanvasWebGL, EngineFimBase, EngineShader } from '@leosingleton/fim/internals';
 import { GlslShader } from 'webpack-glsl-minify';
 
-export class BrowserEngineFim extends EngineFim<BrowserEngineImage, EngineShader> {
+export class BrowserEngineFim extends EngineFimBase<BrowserEngineImage, EngineShader> {
   protected createEngineImage(options: FimImageOptions, dimensions: FimDimensions, imageName?: string):
       BrowserEngineImage {
     return new BrowserEngineImage(this, options, dimensions, imageName);
