+++ packages/fim-node/src/engine/NodeEngineFim.ts
@@ -6,10 +6,10 @@ import { NodeEngineImage } from './NodeEngineImage';
 import { CoreNodeCanvas2D } from '../core/CoreNodeCanvas2D';
 import { CoreNodeCanvasWebGL } from '../core/CoreNodeCanvasWebGL';
 import { FimDimensions, FimImageOptions } from '@leosingleton/fim';
-import { CoreCanvas2D, CoreCanvasOptions, CoreCanvasWebGL, EngineFim, EngineShader } from '@leosingleton/fim/internals';
+import { CoreCanvas2D, CoreCanvasOptions, CoreCanvasWebGL, EngineFimBase, EngineShader } from '@leosingleton/fim/internals';
 import { GlslShader } from 'webpack-glsl-minify';
 
-export class NodeEngineFim extends EngineFim<NodeEngineImage, EngineShader> {
+export class NodeEngineFim extends EngineFimBase<NodeEngineImage, EngineShader> {
   protected createEngineImage(options: FimImageOptions, dimensions: FimDimensions, imageName?: string):
       NodeEngineImage {
     return new NodeEngineImage(this, options, dimensions, imageName);
