+++ packages/fim/src/api/Fim.ts
@@ -16,8 +16,7 @@ import { GlslShader } from 'webpack-glsl-minify';
  * Parent object when using the FIM library. This object holds images, shaders, and operations used to manipulate 2D
  * images within JavaScript using WebGL.
  */
-export interface Fim extends FimBase<FimImage, FimShader> {
-}
+export type Fim = FimBase<FimImage, FimShader>;
 
 /** Templated version of the Fim interface which supports specific implementations of image and shader classes */
 export interface FimBase<TImage extends FimImage, TShader extends FimShader> extends FimObject {
